Fix up transaction types. Add upload functionality. Re gen DB migrations

+++ app/forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, BooleanField, SubmitField, IntegerField, HiddenField
+from wtforms import StringField, PasswordField, BooleanField, SubmitField, DecimalField
 from wtforms.validators import ValidationError, DataRequired, Email, EqualTo
 
 from app.models import User
@@ -46,7 +46,7 @@ class ChildRegistrationForm(FlaskForm):
 
 class AddTransactionForm(FlaskForm):
     # account_id = HiddenField('account_id', validators=[DataRequired()])
-    amount = IntegerField('Amount', validators=[DataRequired()])
+    amount = DecimalField('Amount', validators=[DataRequired()])
     description = StringField('Description', validators=[DataRequired()])
     credit = SubmitField('Credit')
     debit = SubmitField('Debit')
